renamed cofs -> thetis
also removed obsolete install.md
