+++ examples/channel3d/channel3d.py
@@ -10,7 +10,7 @@
 # This test is useful for testing open boundary conditions.
 #
 # Tuomas Karna 2015-03-03
-from cofs import *
+from thetis import *
 
 parameters['coffee'] = {}
