+++ examples/channel3d/channel3d_closed.py
@@ -20,7 +20,7 @@
 # Tuomas Karna 2015-03-03
 
 from scipy.interpolate import interp1d
-from cofs import *
+from thetis import *
 
 n_layers = 6
 outputdir = 'outputs_closed'
