+++ examples/tracerBox/tracerBox3d.py
@@ -13,7 +13,7 @@
 #
 #
 # Tuomas Karna 2015-03-11
-from cofs import *
+from thetis import *
 
 mesh2d = Mesh('channel_wave_eq.msh')
 depth = 50.0
