+++ test/swe2d/test_steady_state_channel.py
@@ -1,5 +1,5 @@
 # Tuomas Karna 2015-03-03
-from cofs import *
+from thetis import *
 import math
