+++ thetis/options.py
@@ -3,7 +3,7 @@ This file defines all options of the 2D/3D models excluding field values.
 
 Tuomas Karna 2015-10-17
 """
-from cofs.utility import *
+from thetis.utility import *
 
 
 class ModelOptions(AttrDict):
