+++ thetis/solver.py
@@ -14,9 +14,9 @@ import time as time_mod
 from mpi4py import MPI
 import exporter
 import weakref
-from cofs.field_defs import field_metadata
-from cofs.options import ModelOptions
-import cofs.callback as callback
+from thetis.field_defs import field_metadata
+from thetis.options import ModelOptions
+import thetis.callback as callback
 
 
 class FlowSolver(FrozenClass):
