+++ thetis/utility.py
@@ -13,8 +13,8 @@ from pyop2.profiling import timed_region, timed_function, timing  # NOQA
 from mpi4py import MPI  # NOQA
 import ufl  # NOQA
 import coffee.base as ast  # NOQA
-from cofs.field_defs import field_metadata
 from collections import OrderedDict  # NOQA
+from thetis.field_defs import field_metadata
 
 comm = op2.MPI.comm
 commrank = op2.MPI.comm.rank
